Rename phase2 -> rn
