+++ rn/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.phase2",
+    package = "com.rn",
 )
 
 android_resource(
     name = "res",
-    package = "com.phase2",
+    package = "com.rn",
     res = "src/main/res",
 )
