+++ rn/android/app/src/main/java/com/rn/MainActivity.java
@@ -1,4 +1,4 @@
-package com.phase2;
+package com.rn;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "phase2";
+    return "rn";
   }
 }
