+++ rn/android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'phase2'
+rootProject.name = 'rn'
 apply from: file("../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesSettingsGradle(settings)
 include ':app'
