+++ rn/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "phase2",
-  "displayName": "phase2"
+  "name": "rn",
+  "displayName": "rn"
 }
