+++ rn/ios/rn/Images.xcassets/AppIcon.appiconset/Contents.json
