+++ rn/ios/rnTests/Info.plist
