+++ rn/ios/rnTests/rnTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface phase2Tests : XCTestCase
+@interface rnTests : XCTestCase
 
 @end
 
-@implementation phase2Tests
+@implementation rnTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
